put the maze, room, player, and item factories in world_builder.py
