+++ dork/game_utils/game_data.py
@@ -1,5 +1,6 @@
 """Data and commands for REPL"""
 
+
 __all__ = ["CMDS", "MOVES", "ERRS", "META", "TITLE"]
 
 
@@ -41,24 +42,19 @@ CMDS = {
     "inv": ["_inventory"],
     "inventory": ["_inventory"],
     "examine": ["_examine"],
-    "grab": ["_take"],
-    "take": ["_take"],
-    "loot": ["_take"],
+    "grab": ["_take_item"],
+    "take": ["_take_item"],
+    "loot": ["_take_item"],
     "use": ["_use_item"],
     "drop": ["_drop_item"],
     "activate": ["_use_item"],
-    # "add": _take,
-    # "loot": _take,
 }
 
 
 META = {
-    ".n": ["_start_over"],
-    ".l": ["_start_over"],
-    ".s": ["_save_game"],
-    ".new": ["_start_over", "new game created!"],
-    ".load": ["_start_over", "game loaded successfully!"],
-    ".save": ["save_game"],
+    ".new": ["_start_over"],
+    ".load": ["_start_over"],
+    ".save": ["_save_game"],
     ".rq": ["_gtfo"],
     ".z": ["_zork"],
     ".m": ["_draw_maze"],
